requirements.txt
python-telegram-bot==21.6
groq>=0.9.0
openpyxl>=3.1.2
gspread>=6.1.2
google-auth>=2.29.0
python-dotenv>=1.0.1
